+++ CHANGELOG.md
@@ -15,6 +15,7 @@
  * **DSL change**: `:on_transition` renamed to `:after`
  * **DSL change**: transition `:after` binding changed (see [issue #59](https://github.com/aasm/aasm/issues/59), thanks to [@stiff](https://github.com/stiff))
  * **DSL change**: instance-based events inspection now returns event instances (instead of the event names as symbol)
+ * **DSL change**: instance-based permissible_events has been removed in favor or events(:permissible => true)
 
 ## 3.9.0 (not yet released)